[tests][s]: moved tests to relevant locations
