+++ test/test_normalize.js
@@ -1,4 +1,5 @@
 const test = require('ava')
+const { data } = require('./data.js')
 const {
   normalizeSchema, normalizeType, nomralizeDateFormat, normalizeAll, normalizeNames
 } = require('../lib/normalize.js')
@@ -220,4 +221,49 @@ test('checks normalized resourse name', t => {
     }]
     }
   t.deepEqual(res, exp)
+})
+
+test('"data help norm[alize]" prints help message for dp command', async t => {
+  const result = await data('help', 'normalize')
+
+  t.is(result.code, 0)
+  const stdout = result.stdout.split('\n')
+  t.true(stdout.length > 1)
+  t.true(stdout[1].includes('Normalize a descriptor (datapackage.json)'))
+})
+
+
+test('"data norm[alize] -h --help" prints help message for dp command', async t => {
+  const result = await data('normalize', '-h')
+
+  t.is(result.code, 0)
+  const stdout = result.stdout.split('\n')
+  t.true(stdout.length > 1)
+  t.true(stdout[1].includes('Normalize a descriptor (datapackage.json)'))
+})
+
+
+test('"data norm[alize] test/fixtures/datapackage.json" normalizes datapackage.json with given file path', async t => {
+  const result = await data('normalize', 'test/fixtures/datapackage.json')
+
+  t.is(result.code, 0)
+  const stdout = result.stdout.split('\n')
+  t.true(stdout.length > 1)
+  t.true(stdout[0].includes('Datapackage.json has been normalized'))
+})
+
+test('"data norm[alize] test/fixtures/" normalizes datapackage.json inside given folder', async t => {
+  const result = await data('normalize', 'test/fixtures/')
+
+  const stdout = result.stdout.split('\n')
+  t.true(stdout[0].includes('Datapackage.json has been normalized'))
+})
+
+test('"data norm[alize] test/fixtures" normalizes datapackage.json inside given folder', async t => {
+  const result = await data('normalize', 'test/fixtures')
+
+  t.is(result.code, 0)
+  const stdout = result.stdout.split('\n')
+  t.true(stdout.length > 1)
+  t.true(stdout[0].includes('Datapackage.json has been normalized'))
 })
